fix: enable cpal on Linux and cast Pulse mic channels to u8
Unblocks Ubuntu release builds: portal-capture can link cpal for mic
listing (parity with Windows), and Pulse Spec.channels matches the
system-audio path. No feature changes.

+++ src-tauri/src/recorder/backend/pulse_audio.rs
@@ -173,7 +173,7 @@ fn run_mic(
 ) {
     let spec = Spec {
         format: Format::F32le,
-        channels: SYSTEM_AUDIO_CHANNELS,
+        channels: SYSTEM_AUDIO_CHANNELS as u8,
         rate: SYSTEM_AUDIO_RATE,
     };
     if !spec.is_valid() {
